pyproject.toml
[tool.poetry]
authors = ["Andrea Mignone <andrea.mignone@virgilio.it>"]
description = "Sync .pre-commit-config.yaml repos starting from poetry.lock"
name = "sync_with_poetry"
version = "1.0.0"

[tool.poetry.dependencies]
python = ">=3.9"

PyYAML = "^6.0.1"
tomlkit = "^0.7.2"

[tool.poetry.group.dev.dependencies]
coverage = [
  {extras = [
    "toml",
  ], version = "^6.0", python = ">=3.6,<3.7"},
  {extras = [
    "toml",
  ], version = ">=6.5", python = ">=3.7"},
]
mypy = "^0.971"
pre-commit = "^3.8.0"
pytest = "^6.2"
pytest-cov = "^4.0.0"
ruff = "^0.6.8"
types-PyYAML = "^6.0.1"

[tool.poetry.scripts]
swp = "sync_with_poetry.swp:main"

[tool.pytest.ini_options]
addopts = "-svv"
testpaths = ["tests"]

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_defs = true
explicit_package_bases = true
files = ["**/*.py"]
namespace_packages = true
show_error_codes = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true
